+++ web/cypress/e2e/incidents/regression/03.reg_api_calls.cy.ts
@@ -29,7 +29,7 @@ const MP = {
 describe('Regression: Silences Not Applied Correctly', { tags: ['@incidents'] }, () => {
 
   before(() => {
-    cy.beforeBlockCOO(MCP, MP);
+    cy.beforeBlockCOO(MCP, MP, { dashboards: false, troubleshootingPanel: false });
   });
 
   beforeEach(() => {
