+++ web/cypress/support/index.ts
@@ -3,6 +3,9 @@ import '@cypress/grep';
 import './selectors';
 import './commands/selector-commands';
 import './commands/auth-commands';
+import './commands/coo-install-commands';
+import './commands/image-patch-commands';
+import './commands/dashboards-commands';
 import './commands/operator-commands';
 import './commands/incident-commands';
 import './commands/utility-commands';
